Pseudonymization_software/requirements.txt
pymupdf==1.23.25
presidio_analyzer==2.2.353
python_docx==1.1.0
spacy==3.7.4
transformers==4.38.1
rapidfuzz==3.6.1
pypdf==4.1.0
pdfminer.six
torch

reportlab~=4.1.0
pillow~=10.2.0
pandas~=2.2.1
numpy~=1.26.4
fuzzywuzzy~=0.18.0
pypdf2~=3.0.1
Faker~=24.4.0
comtypes~=1.4.0
docx2pdf~=0.1.8
doc2docx~=0.2.4